fixed categories CRUD

+++ pages/api/categories/[id].js
@@ -1,17 +1,16 @@
 import connectDB from '../../../utils/connectDB'
-import Categories from '../../../models/categoriesModal'
+import Products from '../../../models/productModel'
 import auth from '../../../middleware/auth'
-
+import Categories from '../../../models/categoriesModel'
 
 connectDB()
 
-
 export default async (req, res) => {
     switch(req.method){
-        case "POST": 
+        case "PUT":
             await updateCategory(req, res)
             break;
-        case "GET": 
+        case "DELETE":
             await deleteCategory(req, res)
             break;
     }
@@ -23,12 +22,12 @@ const updateCategory = async (req, res) => {
         if(result.role !== 'admin')
         return res.status(400).json({err: "Authentication is not valid."})
 
-        const {id} = req.query 
-        const {name} = req.body 
+        const {id} = req.query
+        const {name} = req.body
 
         const newCategory = await Categories.findOneAndUpdate({_id: id}, {name})
         res.json({
-            msg: "Successfully updated category!",
+            msg: "Success! Update a new category",
             category: {
                 ...newCategory._doc,
                 name
@@ -45,11 +44,16 @@ const deleteCategory = async (req, res) => {
         if(result.role !== 'admin')
         return res.status(400).json({err: "Authentication is not valid."})
 
-        const {id} = req.query 
-        const {name} = req.body 
+        const {id} = req.query
+
+        const products = await Products.findOne({category: id})
+        if(products) return res.status(400).json({
+            err: "Please delete all products with a relationship"
+        })
 
         await Categories.findByIdAndDelete(id)
-        res.json({msg: "Successfully deleted a category!"})
+        
+        res.json({msg: "Success! Deleted a category"})
     } catch (err) {
         return res.status(500).json({err: err.message})
     }
